# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ClauseEvaluationInput {
  /**
     * @minLength 10
     * @maxLength 50000
     */
  clauseText: string;
  /**
     * @maxLength 255
     * @nullable
     */
  sourceName?: string | null;
}

export type ClauseEvaluationDecision = typeof ClauseEvaluationDecision[keyof typeof ClauseEvaluationDecision];


export const ClauseEvaluationDecision = {
  INTERCEPT: 'INTERCEPT',
  ALLOW: 'ALLOW',
} as const;

export type ClauseEvaluationRiskLevel = typeof ClauseEvaluationRiskLevel[keyof typeof ClauseEvaluationRiskLevel];


export const ClauseEvaluationRiskLevel = {
  critical: 'critical',
  high: 'high',
  moderate: 'moderate',
  low: 'low',
} as const;

export type AuditEventEventType = typeof AuditEventEventType[keyof typeof AuditEventEventType];


export const AuditEventEventType = {
  CLAUSE_SCREENED: 'CLAUSE_SCREENED',
} as const;

export type AuditEventDecision = typeof AuditEventDecision[keyof typeof AuditEventDecision];


export const AuditEventDecision = {
  INTERCEPT: 'INTERCEPT',
  ALLOW: 'ALLOW',
} as const;

export type AuditEventThreshold = typeof AuditEventThreshold[keyof typeof AuditEventThreshold];


export const AuditEventThreshold = {
  NUMBER_80: 80,
} as const;

export interface AuditEvent {
  auditId: string;
  eventType: AuditEventEventType;
  decision: AuditEventDecision;
  threshold: AuditEventThreshold;
  recordedAt: string;
  actor: string;
}

export interface ClauseEvaluation {
  id: string;
  clauseText: string;
  /** @nullable */
  sourceName: string | null;
  /**
     * @minimum 0
     * @maximum 100
     */
  riskScore: number;
  decision: ClauseEvaluationDecision;
  riskLevel: ClauseEvaluationRiskLevel;
  rationale: string;
  findings: string[];
  provider: string;
  model: string;
  analysisVersion: string;
  createdAt: string;
  audit: AuditEvent;
}

export interface RiskEvaluationSummary {
  total: number;
  intercepted: number;
  allowed: number;
  averageRiskScore: number;
  threshold: number;
  /** @nullable */
  lastEvaluatedAt: string | null;
}

export interface ErrorResponse {
  error: string;
}

export type ListRiskEvaluationsParams = {
/**
 * @minimum 1
 * @maximum 100
 */
limit?: number;
};