# lib/api-zod/src/generated/types/auditEvent.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AuditEventDecision } from './auditEventDecision';
import type { AuditEventEventType } from './auditEventEventType';
import type { AuditEventThreshold } from './auditEventThreshold';

export interface AuditEvent {
  auditId: string;
  eventType: AuditEventEventType;
  decision: AuditEventDecision;
  threshold: AuditEventThreshold;
  recordedAt: Date;
  actor: string;
}