# lib/api-zod/src/generated/types/clauseEvaluation.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AuditEvent } from './auditEvent';
import type { ClauseEvaluationDecision } from './clauseEvaluationDecision';
import type { ClauseEvaluationRiskLevel } from './clauseEvaluationRiskLevel';

export interface ClauseEvaluation {
  id: string;
  clauseText: string;
  /** @nullable */
  sourceName: string | null;
  /**
     * @minimum 0
     * @maximum 100
     */
  riskScore: number;
  decision: ClauseEvaluationDecision;
  riskLevel: ClauseEvaluationRiskLevel;
  rationale: string;
  findings: string[];
  provider: string;
  model: string;
  analysisVersion: string;
  createdAt: Date;
  audit: AuditEvent;
}