# lib/api-zod/src/generated/types/clauseEvaluationInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ClauseEvaluationInput {
  /**
     * @minLength 10
     * @maxLength 50000
     */
  clauseText: string;
  /**
     * @maxLength 255
     * @nullable
     */
  sourceName?: string | null;
}