# lib/api-zod/src/generated/types/riskEvaluationSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RiskEvaluationSummary {
  total: number;
  intercepted: number;
  allowed: number;
  averageRiskScore: number;
  threshold: number;
  /** @nullable */
  lastEvaluatedAt: string | null;
}